PostfixLog/requirements.txt
aiosmtpd==1.4.2
async-timeout==4.0.2
atpublic==3.1.1
attrs==22.2.0
boto3==1.26.48
botocore==1.29.48
certifi==2022.9.24
charset-normalizer==3.0.1
click==8.1.3
Flask==2.2.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.2
prometheus-client==0.15.0
python-dateutil==2.8.2
PyYAML==6.0
redis==4.4.0
reliable-queue==0.0.7
requests==2.28.2
s3transfer==0.6.0
six==1.16.0
types-ujson==4.2.1
ujson==5.7.0
urllib3==1.26.14
Werkzeug==2.2.2
